Fix: 소셜 로그인/회원가입 로직 수정
